+++ src/main/java/ru/syntezis/cronctl/config/swagger/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * OpenAPI / Swagger UI integration configuration.
+ */
 @NullMarked
 package ru.syntezis.cronctl.config.swagger;