+++ src/main/java/ru/syntezis/cronctl/filter/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * Filters that control which {@code @Scheduled} methods are registered.
+ */
 @NullMarked
 package ru.syntezis.cronctl.filter;