+++ src/main/java/ru/syntezis/cronctl/presentation/mapper/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * MapStruct mappers converting domain objects to DTOs.
+ */
 @NullMarked
 package ru.syntezis.cronctl.presentation.mapper;