+++ src/main/java/ru/syntezis/cronctl/processor/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * Processors that extract schedule metadata from annotated bean methods.
+ */
 @NullMarked
 package ru.syntezis.cronctl.processor;