+++ src/main/java/ru/syntezis/cronctl/util/package-info.java
@@ -1,3 +1,6 @@
+/**
+ * Utility classes for cron expression parsing and schedule computation.
+ */
 @NullMarked
 package ru.syntezis.cronctl.util;